requirements.txt
coqui-tts
playsound
nats-py
asyncio
